+++ tests/tde/theme/icons.lua
@@ -294,9 +294,14 @@ function Test_theme_icon_password()
     assert(file_exists(icons.password), "Check that the password icon exists")
 end
 
+function Test_theme_icon_mime()
+    assert(file_exists(icons.mime), "Check that the mime icon exists")
+end
+
+
 
 function Test_that_all_icons_are_tested()
-    local amount = 67
+    local amount = 68
     assert(
         tablelength(icons) == amount,
         "It looks like you forgot to add some icons to the test suite, add them and then update this test (you forgot: " ..